feat: more optimization
