+++ packages/taiko-client-rs/crates/proposer/Cargo.toml
@@ -37,5 +37,5 @@ serde_json = { workspace = true }
 
 [dev-dependencies]
 serial_test = { workspace = true }
-test-harness = { path = "../test-harness" }
 test-context = { workspace = true }
+test-harness = { path = "../test-harness" }
