+++ packages/taiko-client-rs/crates/test-harness/src/lib.rs
@@ -5,10 +5,15 @@ use std::sync::OnceLock;
 use tracing_subscriber::EnvFilter;
 
 mod blob_server;
-mod shasta_env;
+mod helper;
+pub mod shasta;
 
 pub use blob_server::BlobServer;
-pub use shasta_env::{ShastaEnv, mine_l1_block, verify_anchor_block, wait_for_new_proposal};
+pub use helper::mine_l1_block;
+pub use shasta::{
+    env::ShastaEnv,
+    helpers::{verify_anchor_block, wait_for_new_proposal},
+};
 
 /// Initialise tracing for tests using a single global subscriber.
 ///
