+++ packages/taiko-client-rs/crates/test-harness/src/shasta/env.rs
@@ -1,5 +1,4 @@
 use std::{
-    borrow::Cow,
     env, fmt,
     path::PathBuf,
     str::FromStr,
@@ -7,73 +6,23 @@ use std::{
     time::{Duration, Instant},
 };
 
-use alloy::{
-    eips::BlockNumberOrTag, rpc::client::NoParams, sol_types::SolCall,
-    transports::http::reqwest::Url as RpcUrl,
-};
-use alloy_consensus::{Transaction, TxEnvelope};
-use alloy_primitives::{Address, B256, U256};
-use alloy_provider::{
-    Provider, ProviderBuilder, RootProvider, fillers::FillProvider, utils::JoinedRecommendedFillers,
-};
-use alloy_rpc_types::{Transaction as RpcTransaction, eth::Block as RpcBlock};
+use alloy::transports::http::reqwest::Url as RpcUrl;
+use alloy_primitives::{Address, B256};
+use alloy_provider::{ProviderBuilder, RootProvider};
 use anyhow::{Context, Result};
-use bindings::anchor::Anchor::anchorV4Call;
-use event_indexer::indexer::{ProposedEventPayload, ShastaEventIndexer, ShastaEventIndexerConfig};
+use event_indexer::indexer::{ShastaEventIndexer, ShastaEventIndexerConfig};
 use proposer::{config::ProposerConfigs, proposer::Proposer};
 use rpc::{
     SubscriptionSource,
     client::{Client, ClientConfig},
-    error::RpcClientError,
 };
 use test_context::AsyncTestContext;
-use tokio::time::timeout;
-use tracing::warn;
+use tracing::{error, info};
 
-type RpcClient = Client<FillProvider<JoinedRecommendedFillers, RootProvider>>;
-
-const PRECONF_OPERATOR_ACTIVATION_BLOCKS: usize = 64;
-const L1_BLOCK_TIME_SECONDS: u64 = 12;
-
-/// Advances L1 time and mines blocks to ensure the preconfigured operator whitelist is active.
-async fn ensure_preconf_whitelist_active(client: &RpcClient) -> Result<()> {
-    for _ in 0..PRECONF_OPERATOR_ACTIVATION_BLOCKS {
-        increase_l1_time(client, L1_BLOCK_TIME_SECONDS).await?;
-        mine_l1_block(client).await?;
-    }
-    Ok(())
-}
-
-/// Checks if the RPC error indicates a "not found" condition.
-fn is_not_found_error(err: &RpcClientError) -> bool {
-    matches!(err, RpcClientError::Rpc(message) if message.contains("not found"))
-}
-
-/// Reset the authenticated L1 RPC head.
-async fn reset_head_l1_origin(client: &RpcClient) -> Result<()> {
-    match client.set_head_l1_origin(U256::from(1u64)).await {
-        Ok(_) => Ok(()),
-        Err(err) if is_not_found_error(&err) => Ok(()),
-        Err(err) => Err(err.into()),
-    }
-}
-
-/// Revert the L1 snapshot with retries.
-async fn revert_snapshot(provider: &RootProvider, snapshot_id: &str) -> Result<()> {
-    provider
-        .raw_request::<_, bool>(Cow::Borrowed("evm_revert"), (&snapshot_id,))
-        .await
-        .context("reverting L1 snapshot")?;
-    Ok(())
-}
-
-/// Create a new L1 snapshot to reuse across a single test run.
-async fn create_snapshot(phase: &'static str, provider: &RootProvider) -> Result<String> {
-    provider
-        .raw_request::<_, String>(Cow::Borrowed("evm_snapshot"), NoParams::default())
-        .await
-        .with_context(|| format!("creating L1 snapshot during {phase}"))
-}
+use super::helpers::{
+    RpcClient, create_snapshot, ensure_preconf_whitelist_active, reset_head_l1_origin,
+    revert_snapshot,
+};
 
 /// Environment configuration required to exercise Shasta fork integration tests against
 /// the Docker harness started by `tests/entrypoint.sh`.
@@ -89,7 +38,7 @@ pub struct ShastaEnv {
     pub l1_proposer_private_key: B256,
     pub taiko_anchor_address: Address,
     pub client_config: ClientConfig,
-    pub client: Client<FillProvider<JoinedRecommendedFillers, RootProvider>>,
+    pub client: RpcClient,
     pub event_indexer: Arc<ShastaEventIndexer>,
     pub proposer: Arc<Proposer>,
     cleanup_provider: RootProvider,
@@ -97,6 +46,7 @@ pub struct ShastaEnv {
 }
 
 impl fmt::Debug for ShastaEnv {
+    /// Formats the `ShastaEnv` for debugging, omitting sensitive fields.
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("ShastaEnv")
             .field("l1_source", &self.l1_source)
@@ -109,67 +59,8 @@ impl fmt::Debug for ShastaEnv {
             .field("taiko_anchor_address", &self.taiko_anchor_address)
             .field("client_config", &self.client_config)
             .field("client", &self.client)
-            .field("event_indexer", &"ShastaEventIndexer")
-            .field("proposer", &"Proposer")
             .finish()
     }
-}
-
-/// Ensures the latest L2 block contains an Anchor `anchorV4` call.
-pub async fn verify_anchor_block<P>(client: &Client<P>, anchor_address: Address) -> Result<()>
-where
-    P: alloy_provider::Provider + Clone + Send + Sync + 'static,
-{
-    let latest_block: RpcBlock<TxEnvelope> = client
-        .l2_provider
-        .get_block_by_number(BlockNumberOrTag::Latest)
-        .full()
-        .await?
-        .ok_or_else(|| anyhow::anyhow!("latest block missing"))?
-        .map_transactions(|tx: RpcTransaction| tx.into());
-
-    let first_tx = latest_block
-        .transactions
-        .as_transactions()
-        .and_then(|txs| txs.first())
-        .ok_or_else(|| anyhow::anyhow!("block missing anchor transaction"))?;
-
-    let selectors = [anchorV4Call::SELECTOR];
-    anyhow::ensure!(first_tx.input().len() >= 4, "anchor transaction input too short");
-    anyhow::ensure!(
-        selectors.iter().any(|sel| &first_tx.input()[..sel.len()] == sel.as_slice()),
-        "first transaction is not calling an Anchor anchorV4 entrypoint"
-    );
-    anyhow::ensure!(
-        first_tx.to() == Some(anchor_address),
-        "anchor transaction target mismatch: expected {}, got {:?}",
-        anchor_address,
-        first_tx.to()
-    );
-
-    Ok(())
-}
-
-/// Waits until the indexer observes a proposal with an ID greater than `previous_id`.
-pub async fn wait_for_new_proposal(
-    indexer: Arc<ShastaEventIndexer>,
-    previous_id: u64,
-) -> Result<ProposedEventPayload> {
-    let wait = async {
-        loop {
-            if let Some(payload) = indexer.get_last_proposal() {
-                let proposal_id = payload.proposal.id.to::<u64>();
-                if proposal_id > previous_id {
-                    return payload;
-                }
-            }
-            tokio::time::sleep(Duration::from_millis(100)).await;
-        }
-    };
-
-    timeout(Duration::from_secs(15), wait)
-        .await
-        .map_err(|_| anyhow::anyhow!("timed out waiting for proposal to be indexed"))
 }
 
 impl ShastaEnv {
@@ -248,7 +139,7 @@ impl ShastaEnv {
         let proposer =
             Arc::new(Proposer::new_with_indexer(proposer_config, event_indexer.clone()).await?);
 
-        tracing::info!(elapsed_ms = started.elapsed().as_millis(), "loaded ShastaEnv");
+        info!(elapsed_ms = started.elapsed().as_millis(), "loaded ShastaEnv");
         Ok(Self {
             l1_source,
             l2_http: l2_http_url,
@@ -282,30 +173,7 @@ impl AsyncTestContext for ShastaEnv {
     /// Teardown the ShastaEnv after each test.
     async fn teardown(self) {
         if let Err(err) = self.shutdown().await {
-            warn!(error = %err, "failed to teardown ShastaEnv");
+            error!(error = %err, "failed to teardown ShastaEnv");
         }
     }
-}
-
-/// Mines a single empty L1 block via the connected execution engine.
-pub async fn mine_l1_block<P>(client: &Client<P>) -> Result<()>
-where
-    P: Provider + Clone + Send + Sync + 'static,
-{
-    client
-        .l1_provider
-        .raw_request::<_, String>(Cow::Borrowed("evm_mine"), NoParams::default())
-        .await
-        .context("mining L1 block")?;
-    Ok(())
-}
-
-/// Increases L1 time by the specified number of seconds.
-async fn increase_l1_time(client: &RpcClient, seconds: u64) -> Result<()> {
-    client
-        .l1_provider
-        .raw_request::<_, i64>(Cow::Borrowed("evm_increaseTime"), (seconds,))
-        .await
-        .context("increasing L1 time via evm_increaseTime")?;
-    Ok(())
 }
