+++ packages/taiko-client-rs/crates/test-harness/src/shasta/mod.rs
@@ -1,0 +1,5 @@
+pub mod env;
+pub mod helpers;
+
+pub use env::ShastaEnv;
+pub use helpers::{verify_anchor_block, wait_for_new_proposal};
